move registration hooks (doesn't compile)
