+++ mlir/lib/TableGen/OpDocGen.cpp
@@ -603,42 +603,3 @@ static bool emitDialectDoc(const RecordKeeper &records, raw_ostream &os) {
                  dialectEnums, os);
   return false;
 }
-
-//===----------------------------------------------------------------------===//
-// Gen Registration
-//===----------------------------------------------------------------------===//
-
-static mlir::GenRegistration
-    genAttrDocRegister("gen-attrdef-doc",
-                       "Generate dialect attribute documentation",
-                       [](const RecordKeeper &records, raw_ostream &os) {
-                         emitAttrOrTypeDefDoc(records, os, "AttrDef");
-                         return false;
-                       });
-
-static mlir::GenRegistration
-    genOpDocRegister("gen-op-doc", "Generate dialect documentation",
-                     [](const RecordKeeper &records, raw_ostream &os) {
-                       emitOpDoc(records, os);
-                       return false;
-                     });
-
-static mlir::GenRegistration
-    genTypeDocRegister("gen-typedef-doc", "Generate dialect type documentation",
-                       [](const RecordKeeper &records, raw_ostream &os) {
-                         emitAttrOrTypeDefDoc(records, os, "TypeDef");
-                         return false;
-                       });
-
-static mlir::GenRegistration
-    genEnumDocRegister("gen-enum-doc", "Generate dialect enum documentation",
-                       [](const RecordKeeper &records, raw_ostream &os) {
-                         emitEnumDoc(records, os);
-                         return false;
-                       });
-
-static mlir::GenRegistration
-    genPassDocRegister("gen-dialect-doc", "Generate dialect documentation",
-                   [](const RecordKeeper &records, raw_ostream &os) {
-                     return emitDialectDoc(records, os);
-                   });
